pyproject.toml
[project]
name = "metu-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.53.1",
    "tqdm>=4.67.1",
]
